Update BluetoothService to use custom hooks
